Add separate build and develop commands, and more
- specifying "default" is now optional
- more tests

+++ flake.nix
@@ -7,40 +7,78 @@
     with builtins;
     inputs.utils.apply-systems { inherit inputs; }
       ({ pkgs, ... }:
-         let
-           l = p.lib; p = pkgs;
-           test = { n }: p.runCommand "test" {} "echo ${toString n} > $out";
-         in
+         let l = p.lib; p = pkgs; in
          rec
          { packages =
              { default =
                  p.writeShellScriptBin "nargs"
                    ''
-                   tmp=$(mktemp)
+                   if [[ "$1" == build || "$1" == develop ]]; then
+                     if [[ "$1" == build ]]; then
+                       type=packages
+                     else
+                       type=devShells
+                     fi
 
-                   echo "with builtins;
-                   { ... }@args:
-                     (getFlake \"path:$PWD\")
-                       .packages
-                       .\''${currentSystem}
-                       .\"$1\"
-                       .override (_: args)" > $tmp
+                     if [[ "$2" == -* || "$2" == "" ]]; then
+                       package=default
+                       argstart=2
+                     else
+                       package="$2"
+                       argstart=3
+                     fi
 
-                   nix build -f $tmp "''${@:2}"
-                   rm $tmp
+                     tmp=$(mktemp)
+
+                     echo "with builtins;
+                     { ... }@args:
+                       let
+                         package =
+                           (getFlake \"path:$PWD\")
+                             ."''${type}"
+                             .\''${currentSystem}
+                             .\"$package\";
+                       in
+                       if package?override
+                       then package.override (_: args)
+                       else package" > $tmp
+
+                     nix "$1" -f $tmp "''${@:$argstart}"
+                     rm $tmp
+                   else
+                     echo "nargs build [package-name] [nix args]"
+                     echo "nargs develop [package-name] [nix args]"
+                   fi
                    '';
 
-               test = p.callPackage test { n = 0; };
+               test =
+                 let
+                   test = { n }: p.runCommand "test" {} "echo ${toString n} > $out";
+                 in
+                 p.callPackage test { n = 0; };
              };
+
+           devShells.test =
+             let
+               test = { n }:
+                 p.mkShell
+                   { buildInputs =
+                       [ (p.writeShellScriptBin "test-script" "echo ${toString n}") ];
+                   };
+             in
+             p.callPackage test { n = 0; };
 
            apps.test =
              { type = "app";
 
                program =
+                 let nargs = "${packages.default}/bin/nargs"; in
                  (p.writeShellScript "test-script"
                     ''
-                    ${packages.default}/bin/nargs test -o test --arg n 1
-                    if [[ $(cat test) == 1 ]]; then
+                    if [[ $(cat $(${nargs} build test --no-link --print-out-paths --arg n 1)) == 1
+                       && $(${nargs} develop test --arg n 1 --command test-script) == 1
+                       ]]
+                    then
                       echo test passed
                     else
                       echo test failed
